pyproject.toml
[project]
name = "py-fractals"
version = "0.1.0"
description = ""
authors = [
    {name = "Eugene Akimov",email = "hyperflex19@gmail.com"}
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11, <3.14"
dependencies = [
    "pyside6 (>=6.8.2.1,<7.0.0.0)",
    "pyopengl (>=3.1.9,<4.0.0)",
    "pyopengl-accelerate (>=3.1.9,<4.0.0)",
    "ffmpeg (>=1.4,<2.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)"
]

[tool.poetry]
packages = [
    {include = "app", from = "src"},
    {include = "fractals", from = "src"},
    {include = "frontend", from = "src"},
    {include = "util", from = "src"},
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120